Change validator to use URI.parse over Addressable::URI.parse
In Guide we use `URI.parse` to parse urls in many places. However, this
validator gem uses `Addressable::URI.parse` to do url validation which
allows for many non-utf8 characters that cause an error when passed into
`URI.parse`.

This PR changes the validator to use `URI.parse` for consistency and to
avoid urls passing validation that then fail when calling `URI.parse` on
the string later in the view layer.

e.g. `URI.parse('http://www.中国政府.政务.cn')` fails, however
`Addressable::URI.parse('http://www.中国政府.政务.cn')` does not error.

Since we currently do not need to support international characters in
URLs we should use `URI.parse` since that is already being used
throughout the guide codebase. If in the future we want to support
international characters we should revisit and potentially change all
`URI.parse` calls to `Addressable::URI.parse` in this gem and in the
guide codebase.